+++ src/hooks/useResolveMenuBar.tsx
@@ -1,7 +1,7 @@
 import useResolution, { MenuBar } from './useDetectResolution';
 import { useRouter } from 'next/router';
 
-import WritePageMenus from 'src/components/navigation/WritePageMenus/WritePageMenus';
+import WritePageMenu from 'src/components/navigation/write-page-menu/WritePageMenu';
 import Menus from 'src/components/navigation/Menus/Menus';
 import MobileMenu from 'src/components/navigation/mobile-menu/MobileMenu';
 
@@ -10,7 +10,7 @@ export const useResolveMenuBar = () => {
   const router = useRouter();
 
   const resolveMenuBar = () => {
-    if (router.pathname.includes('/write')) return <WritePageMenus />;
+    if (router.pathname.includes('/write')) return <WritePageMenu />;
     else return <Menus />;
   };
